refactor(direct-control): extract session socket helper
Refs: openspec/changes/civ7-support-direct-control-modularization

Moves openCiv7TunerSocket into src/session/socket.ts while keeping Civ7DirectControlSession, frame request handling, state querying, command execution, session creation/close, reconnect orchestration, and public facade exports in index.ts.

Preserves connection timeout and connection-failed error classification. This does not define telemetry/correlation contracts, public procedure schemas, semantic CLI projection, AI ingestion, hotseat runtime proof, or Effect/oRPC procedure-core work. Task 2.9.4 matrix-row acceptance remains pending.

Validation:
- bun run --cwd packages/civ7-direct-control test -- session.test.ts public-api.test.ts
- bun run --cwd packages/civ7-direct-control check
- bun run --cwd packages/civ7-direct-control test
- bun run --cwd packages/civ7-direct-control build
- bun run openspec -- validate civ7-support-direct-control-modularization --strict
- bun run resources:status
- bun run check:cli
- bun run test:cli:play
- bun run --cwd apps/mapgen-studio check
- git diff --check
- git diff --cached --check
- ownership and no-cycle scans for session socket helper

Local package/source relocation proof only; no runtime, live-game, AI-on-hotseat, telemetry, semantic CLI, or procedure-core proof claimed.
