+++ openspec/changes/civ7-support-direct-control-modularization/tasks.md
@@ -582,6 +582,13 @@ runtime/direct-control claims.
         telemetry/correlation contracts, procedure schemas, semantic CLI, AI
         ingestion, hotseat runtime proof, Effect/oRPC procedure-core work, and
         Task 2.9.4 matrix-row acceptance pending.
+  - [x] 4.10.7 Extract session socket-open helper owner while preserving
+        connection timeout and connection-failed error classification and
+        leaving `Civ7DirectControlSession`, frame request handling, state
+        querying, command execution, reconnect orchestration,
+        telemetry/correlation contracts, procedure schemas, semantic CLI, AI
+        ingestion, hotseat runtime proof, Effect/oRPC procedure-core work, and
+        Task 2.9.4 matrix-row acceptance pending.
 - [x] 4.11 Extract map/visibility/GameInfo read atoms.
   - [x] 4.11.1 Extract map summary, plot snapshot, and map grid read
         wrapper/source owner while keeping the public facade export surface in
