+++ packages/civ7-direct-control/src/session/socket.ts
@@ -1,0 +1,36 @@
+import { Socket, createConnection } from "node:net";
+
+import { Civ7DirectControlError } from "../direct-control-error.js";
+
+export async function openCiv7TunerSocket(options: {
+  host: string;
+  port: number;
+  timeoutMs: number;
+}): Promise<Socket> {
+  return await new Promise<Socket>((resolve, reject) => {
+    const socket = createConnection({ host: options.host, port: options.port });
+    const timer = setTimeout(() => {
+      socket.destroy();
+      reject(
+        new Civ7DirectControlError(
+          "connection-timeout",
+          `Timed out connecting to Civ7 tuner socket ${options.host}:${options.port}`,
+        ),
+      );
+    }, options.timeoutMs);
+    socket.once("connect", () => {
+      clearTimeout(timer);
+      resolve(socket);
+    });
+    socket.once("error", (err) => {
+      clearTimeout(timer);
+      reject(
+        new Civ7DirectControlError(
+          "connection-failed",
+          `Failed connecting to Civ7 tuner socket ${options.host}:${options.port}: ${err.message}`,
+          { cause: err },
+        ),
+      );
+    });
+  });
+}
